feat(ui): look and feel wip
settings dialog updates (wip)
