+++ src/renderer/style2/imports.css
@@ -1,4 +1,5 @@
 @import "tabs.css";
+@import "forms.css";
 @import "icons.css";
 @import "starting-engine.css";
 @import "new-job.css";
@@ -6,3 +7,4 @@
 @import "drop-target.css";
 @import "script-form.css";
 @import "job-details.css";
+@import "app-settings.css";
